backend/requirements.txt
# Web框架
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Mail==0.9.1

# 数据库
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
alembic==1.12.0

# 任务队列
celery==5.3.6
redis==5.0.0
flower==2.0.1

# Ansible集成
ansible==8.5.0
ansible-core==2.15.5
ansible-runner==2.3.4
# ansible-vault包已包含在ansible中

# HTTP客户端
requests==2.31.0
httpx==0.25.0

# 数据处理
PyYAML==6.0.1
jsonschema==4.19.1
marshmallow==3.20.1

# 日期时间
python-dateutil==2.8.2
pytz==2023.3

# 文件处理
Pillow==10.0.1
openpyxl==3.1.2

# 开发工具
python-dotenv==1.0.0
click==8.1.7

# 工具库
itsdangerous==2.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
Jinja2==3.1.2

# 认证和安全
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.4
passlib==1.7.4
pyotp==2.9.0
qrcode==7.4.2

# 监控和日志
prometheus-client==0.17.1
structlog==23.1.0
coloredlogs==15.0.1
psutil==5.9.5

# 部署
gunicorn==21.2.0
gevent==23.7.0
eventlet==0.33.3

# 测试
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.6.2

# 代码质量
flake8==6.1.0
black==23.9.1
isort==5.12.0
mypy==1.5.1

# 文档
flasgger==0.9.7.1

# 工具插件
click-plugins==1.1.1